+++ .formatter.exs
@@ -1,3 +1,9 @@
 [
+  locals_without_parens: [
+    defquery: 2,
+    defquery: 3,
+    definput: 1,
+    definput: 2
+  ],
   inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
 ]