+++ test/support/empty_queries.ex
@@ -1,0 +1,4 @@
+defmodule Roux.Test.EmptyQueries do
+  @moduledoc false
+  use Roux.Query
+end